requirements.txt
fastapi
uvicorn
streamlit
langchain==0.1.11
langchain-openai==0.0.5
openai==1.12.0
requests
python-dotenv
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
httpx==0.25.2
pydantic==2.5.0
google-generativeai

